Seller UI
TODO: Bluteooth Connection
