+++ app/src/main/java/com/app/brensurio/iorder/activities/MainActivity.java
@@ -113,6 +113,8 @@ public class MainActivity extends AppCompatActivity implements MyFragmentCallbac
                                             intent.putExtra(MainActivity.STORE_NAME,
                                                     MainActivity.STORE_3);
                                         //loading();
+                                        intent.putExtra("NAME", name);
+                                        intent.putExtra("EMAIL", email);
                                         startActivity(intent);
                                         finish();
                                     } else if (eid.substring(0, 1).equalsIgnoreCase("f")) {
