+++ app/src/main/java/com/app/brensurio/iorder/adapters/SellerOrderAdapter.java
@@ -1,7 +1,10 @@
 package com.app.brensurio.iorder.adapters;
 
+import android.bluetooth.BluetoothAdapter;
 import android.content.Context;
+import android.content.Intent;
 import android.graphics.drawable.Drawable;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.CardView;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -12,13 +15,10 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.app.brensurio.iorder.R;
+import com.app.brensurio.iorder.activities.SellerOrderDetailActivity;
 import com.app.brensurio.iorder.models.Order;
 
 import java.util.List;
-
-/**
- * Created by Mariz L. Maas on 11/14/2016.
- */
 
 public class SellerOrderAdapter extends
         RecyclerView.Adapter<SellerOrderAdapter.ViewHolder> {
@@ -28,7 +28,6 @@ public class SellerOrderAdapter extends
     private Listener listener;
 
     public interface Listener {
-        void onViewOrder(int position);
         void onClickDeleteItem(int position);
     }
 
@@ -54,11 +53,15 @@ public class SellerOrderAdapter extends
     }
 
     @Override
-    public void onBindViewHolder(ViewHolder holder, final int position) {
+    public void onBindViewHolder(final ViewHolder holder, int position) {
         CardView cardView = holder.cardView;
 
         TextView refTextView = (TextView) cardView.findViewById(R.id.ref_text_view);
         refTextView.setText("Order ID: " + orders.get(position).getRefNo().substring(6, 12));
+
+        TextView dateTextview = (TextView) cardView.findViewById(R.id.date_text_view);
+        dateTextview.setText(orders.get(position).getDatetime());
+
         TextView statusTextView = (TextView) cardView.findViewById(R.id.status_text_view);
         statusTextView.setText(orders.get(position).getStatus());
 
@@ -81,9 +84,10 @@ public class SellerOrderAdapter extends
         viewOrderButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                if (listener != null) {
-                    listener.onViewOrder(position);
-                }
+                Intent intent = new Intent(context, SellerOrderDetailActivity.class);
+                intent.putExtra("order", orders.get(holder.getAdapterPosition()));
+                intent.putExtra("STORE_NAME", orders.get(holder.getAdapterPosition()).getStore());
+                context.startActivity(intent);
             }
         });
 
@@ -92,9 +96,9 @@ public class SellerOrderAdapter extends
             @Override
             public void onClick(View view) {
                 if (listener != null) {
-                    listener.onClickDeleteItem(position);
-                    notifyItemRemoved(position);
-                    notifyItemRangeChanged(position, getItemCount());
+                    listener.onClickDeleteItem(holder.getAdapterPosition());
+                    notifyItemRemoved(holder.getAdapterPosition());
+                    notifyItemRangeChanged(holder.getAdapterPosition(), getItemCount());
                 }
             }
         });
