+++ app/src/main/java/com/app/brensurio/iorder/fragments/SellerOrderFragment.java
@@ -30,15 +30,11 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-/**
- * A simple {@link Fragment} subclass.
- */
 public class SellerOrderFragment extends Fragment {
 
     private List<Order> orderList;
     private DatabaseReference mDatabase;
     private SellerFragmentListener sellerFragmentListener;
-
 
     public SellerOrderFragment() {
         // Required empty public constructor
@@ -77,49 +73,6 @@ public class SellerOrderFragment extends Fragment {
                 LinearLayoutManager linearLayoutManager =
                         new LinearLayoutManager(getActivity());
                 sellerOrderAdapter.setListener(new SellerOrderAdapter.Listener() {
-                    @Override
-                    public void onViewOrder(int position) {
-                        // total | date | refno | location | subtotal1 | foodname1 | quantity1 | subtotalN | foodnameN | quantityN
-                        String data;
-
-
-                        String[] name =
-                                orderList.get(position).getCustomerName().split("\\s+");
-                        int last = name.length - 1;
-
-                        String total = Double.toString(orderList.get(position).getAmount());
-                        String date = orderList.get(position).getDatetime().substring(0, 17);
-                        Toast.makeText(getActivity(), date, Toast.LENGTH_SHORT).show();
-                        String refNo = orderList.get(position).getRefNo().substring(6, 12);
-                        String location = orderList.get(position).getLocation();
-
-                        data = name[0] + "|" + name[last] + "|" + total + "|" + date + "|" + refNo + "|" + location + "|";
-
-                        for (Food food : orderList.get(position).getItems()) {
-                            String subtotal;
-                            String price;
-                            String foodname;
-                            String quantity;
-
-                            subtotal = Double.toString(food.getPrice() * food.getAmount());
-                            price = Double.toString(food.getPrice());
-                            foodname = food.getName();
-                            quantity = Integer.toString(food.getAmount());
-
-                            data += subtotal + "|" + price + "|" + quantity + "|" + foodname + "|";
-                        }
-                        sellerFragmentListener.setString(data);
-                        sellerFragmentListener.transmitData();
-
-                        DatabaseReference databaseReference =
-                                mDatabase.child("storeorders")
-                                        .child(orderList.get(position).getRefNo());
-                        Map<String, Object> statusUpdate = new HashMap<>();
-                        statusUpdate.put("status", "CONFIRMED");
-                        databaseReference.updateChildren(statusUpdate);
-
-                    }
-
                     @Override
                     public void onClickDeleteItem(int position) {
                         DatabaseReference databaseReference =
