+++ app/src/main/java/com/app/brensurio/iorder/interfaces/SellerFragmentListener.java
@@ -1,11 +1,5 @@
 package com.app.brensurio.iorder.interfaces;
 
-/**
- * Created by Mariz L. Maas on 11/13/2016.
- */
-
 public interface SellerFragmentListener {
     String getStoreName();
-    void setString(String s);
-    void transmitData();
 }
