+++ app/src/main/res/layout/activity_seller_order_detail.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/activity_seller_order_detail"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="com.app.brensurio.iorder.activities.SellerOrderDetailActivity">
+
+    <RelativeLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_alignParentTop="true"
+        android:layout_alignParentStart="true"
+        android:layout_above="@+id/print_button"
+        android:id="@+id/placeholder">
+
+    </RelativeLayout>
+
+    <Button
+        android:text="Button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:id="@+id/print_button"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentStart="true"
+        style="@style/Widget.AppCompat.Button.Colored" />
+
+</RelativeLayout>
